requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
openai==1.58.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
ephem==4.1.6
python-dotenv==1.0.1
numpy<2.0
